backend/requirements.txt
asgiref==3.5.0

boto3==1.21.6
botocore==1.24.6
Django
django-ckeditor==5.9.0
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-js-asset==1.2.2
django-storages==1.12.3
djangorestframework==3.11.1
gunicorn==20.1.0
jmespath==0.10.0
Pillow==9.0.1
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
